Add new module 'kafka_topics'
The kafka_topics module creates new topics in Kafka
(http://kafka.apache.org/), or modifies existing ones (though only
increasing the partition count is supported in Kafka).

Topics can be operated on one at a time, or in a group, to save on
starting up the JVM for each topic to check its current state.